requirements.txt
amqp==5.0.2
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1.3
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-js-asset==1.2.2
django-redis==4.12.1
djangorestframework==3.12.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
openapi-codec==1.3.2
pandas==1.2.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.7.0
Pygments==2.7.3
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
redis==3.5.3
requests==2.25.0
sentry-sdk==0.19.5
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
